pyproject.toml
[tool.poetry]
name = "kurs5"
version = "0.1.0"
description = ""
authors = ["Ilia Akatov <ilia-akatoff@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
poetry-plugin-dotenv = "^2.1.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
